Fix Dialyzer run failing in CI tests
Dialyzer failed in CI during testing because we'd run the tests in a Docker image which uses the output of a build of OTP, but that build doesn't include the `scripts/` directory from the main Git repository. This change mirrors that directory into the Docker image used for testing so that it can be run there.